refactor(airflow): switch to external calitp library
