+++ airflow/requirements.txt
@@ -6,3 +6,4 @@ gusty==0.6.0
 pandas-gbq==0.14.1
 pybigquery==0.7.0
 google-cloud-bigquery==2.16.1
+calitp==0.0.1
